SubGhz: fold the allocs and deserializes the earlier reviews left behind
Follow-up to 84aca75eb and 5f4f4cc2b, both of which deliberately stopped at the
mechanical cases and listed what they skipped.

15 more allocs now call the shared helpers and keep only their own extra work -
keystore lookups, rainbow-table names, extra buffers: alutech_at_4n, beninca_arc,
bin_raw, dickert_mahs, faac_slh, jarolift, keeloq, kinggates_stylo_4k,
nice_flor_s, telcoma_edge.

8 more encoders join subghz_protocol_encoder_common_deserialize. Their
get_upload had no failure path and returned void, so it becomes
bool/void* and returns true; five of them also move their
check_remote_controller call from deserialize into the top of get_upload,
which needs a forward declaration since get_upload is defined above it.

min_count_bit widens from uint8_t to uint16_t in both deserialize helpers.
subghz_block_generic_deserialize_check_count_bit compares it against a
uint16_t, so the narrower parameter would have silently truncated any protocol
above 255 bits. None exists today; the tree's largest is 128.

The get_upload doc says "true Always; this encoder has no failure path" rather
than "on success". These eight have no bound check, unlike the eleven converted
in 5f4f4cc2b, and claiming success would imply a validation none of them do.
They cannot overrun today - deserialize_check_count_bit pins data_count_bit by
equality before get_upload runs - but keyfinder has only 4 spare entries in a
60-entry buffer, so adding the guards is worth its own commit.

Not done here, deliberately: the five moved check_remote_controller calls write
only serial/btn/cnt on the encoder instance, which nothing reads, so they could
be deleted outright rather than moved. Deleting a store is a behaviour change
and belongs with the encoder.front reset in its own commit.

.text 692,296 -> 691,848 (-448 B), .rodata unchanged, api_symbols.csv untouched.

+++ lib/subghz/protocols/common.c
@@ -132,7 +132,7 @@ void* subghz_protocol_encoder_common_alloc(
 SubGhzProtocolStatus subghz_protocol_encoder_common_deserialize(
     void* context,
     FlipperFormat* flipper_format,
-    uint8_t min_count_bit,
+    uint16_t min_count_bit,
     SubGhzProtocolEncoderGetUpload get_upload) {
     furi_assert(context);
     SubGhzProtocolEncoderCommonGeneric* instance = context;
@@ -155,7 +155,7 @@ SubGhzProtocolStatus subghz_protocol_encoder_common_deserialize(
 SubGhzProtocolStatus subghz_protocol_decoder_common_deserialize_te(
     void* context,
     FlipperFormat* flipper_format,
-    uint8_t min_count_bit) {
+    uint16_t min_count_bit) {
     furi_assert(context);
     SubGhzProtocolDecoderCommonTe* instance = context;
     SubGhzProtocolStatus ret = subghz_block_generic_deserialize_check_count_bit(
